% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NLint}
\alias{NLint}
\title{Estimate nonlinear association and identify interactions}
\usage{
NLint(Y = Y, X = X, C = C, nChains = 2, nIter = 10000, nBurn = 2000,
  thin = 8, c = 0.001, d = 0.001, sigB = "EB", k = 15, ns = 3,
  alph = 3, gamm = dim(X)[2], probSamp1 = 0.5, threshold = 0.25)
}
\arguments{
\item{c}{An n by q matrix of additional covariates to adjust for}

\item{y}{The outcome to be analyzed}

\item{x}{an n by p matrix of exposures to evaluate}
}
\value{
An list of values that contain the treatment effect, confidence interval for the 
        treatment effect, and full posterior draws for the treatment effect.
}
\description{
This function takes in the observed data (y, x, c) and estimates
a potentially nonlinear, interactive effect between x and y while
adjusting for c linearly
}
\examples{

n = 200
p = 10
pc = 1

sigma = matrix(0.3, p, p)
diag(sigma) = 1
X = rmvnorm(n, mean=rep(0,p), sigma = sigma)

C = matrix(rnorm(n*pc), nrow=n)

TrueH = function(X) {
  return(0.5*(X[,2]*X[,3]) - 0.6*(X[,4]^2 * X[,5]))
}

Y = 5 + C + TrueH(X) + rnorm(n)

NLmod = NLint(Y=Y, X=X, C=C)

## Print posterior inclusion probabilities
NLmod$MainPIP

## Show the two way interaction probability matrix
NLmod$InteractionPIP

}

