% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NLpredict}
\alias{NLpredict}
\title{Predict outcome at a given set of locations}
\usage{
NLpredict(NLmod = NLmod, X = X, Xnew = Xnew, Cnew = Cnew)
}
\arguments{
\item{NLmod}{A model built from the NLint function that predictions will be based on}

\item{X}{The original matrix of exposures used to build NLmod}

\item{Xnew}{The new set of exposures at which predictions are to be made for}

\item{Cnew}{Matrix of additional covariates at which to make predictions}
}
\value{
The posterior means, 95% pointwise credible intervals, and full posterior draws
        for the predicted outcome
}
\description{
This function takes in new data points and estimates
the posterior predictive distribution of outcome at these
locations
}
\examples{

n = 200
p = 10
pc = 1

sigma = matrix(0.3, p, p)
diag(sigma) = 1
X = rmvnorm(n, mean=rep(0,p), sigma = sigma)

C = matrix(rnorm(n*pc), nrow=n)

TrueH = function(X) {
  return(0.5*(X[,2]*X[,3]) - 0.6*(X[,4]^2 * X[,5]))
}

Y = 5 + C + TrueH(X) + rnorm(n)

NLmod = NLint(Y=Y, X=X, C=C)

Xnew = matrix(rnorm(200*p), 200, p)
Cnew = rnorm(200)

predictions = NLpredict(NLmod=NLmod,
                        X=X, Xnew=Xnew,
                        Cnew=Cnew)
}
