% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NLint}
\alias{NLint}
\title{Estimate nonlinear association and identify interactions}
\usage{
NLint(Y = Y, X = X, C = C, nChains = 2, nIter = 10000, nBurn = 2000,
  thin = 8, c = 0.001, d = 0.001, sigB = "EB", k = 15, ns = 3,
  alph = 3, gamm = dim(X)[2], probSamp1 = 0.9, threshold = 0.1)
}
\arguments{
\item{Y}{The outcome to be analyzed.}

\item{X}{an n by p matrix of exposures to evaluate. These should be continuous.}

\item{C}{An n by q matrix of additional covariates to adjust for.}

\item{nBurn}{The number of MCMC scans that will be dropped as a burn-in.}

\item{thin}{This number represents how many iterations between each scan.}

\item{c}{The first hyperparameter of the inverse gamma prior on the residual variance}

\item{d}{The second hyperparameter of the inverse gamma prior on the residual variance}

\item{sigB}{Either a numeric value to be used for the value of sigma_beta,
the variance of the slab distribution,
or "EB" is specified to indicate that it will be estimated
via empirical Bayes. We recommend "EB" unless the user has
strong prior beliefs about the magnitude of the nonzero regression
coefficients}

\item{k}{The number of total components to allow in the model}

\item{ns}{The degrees of freedom of the splines used to estimate nonlinear functions of the exposures}

\item{alph}{The first hyperparameter for the beta prior on the probability of an exposure being included
in a given component}

\item{gamm}{The second hyperparameter for the beta prior on the probability of an exposure being included
in a given component}

\item{probSamp1}{The proportion of the time the MCMC updates one variable at a time for each component as opposed to
two at a time. We have found that sampling one at a time is faster computationally, but we don't 
recommend probSamp1 = 1, because it is possible for some interactions to be significant without 
their main effects being significant, and these associations are much easier to pick up 2 at a time}

\item{threshold}{thresholding parameter when finding the lower bound for the slab variance. This parameter represents
the percentage of time a null association enters the model when tau_h = 0.5. Smaller values are more
conservative and prevent false discoveries. We recommend either 0.25 or 0.1. This lower bound is only
calculated if the "EB" option is selected, as it is used to make sure the empirical Bayes variance isn't
too small}

\item{nScans}{The number of MCMC scans to run.}
}
\value{
A list containing the full posterior draws of all parameters in the model, 
        the waic associated with the model, the posterior inclusion probabilities (PIPs)
        for each exposure entering into the model, and the matrix of 2-way interaction
        probabilities
}
\description{
This function takes in the observed data (y, x, c) and estimates
a potentially nonlinear, interactive effect between x and y while
adjusting for c linearly
}
\examples{

n = 200
p = 10
pc = 1

sigma = matrix(0.3, p, p)
diag(sigma) = 1
X = rmvnorm(n, mean=rep(0,p), sigma = sigma)

C = matrix(rnorm(n*pc), nrow=n)

TrueH = function(X) {
  return(0.5*(X[,2]*X[,3]) - 0.6*(X[,4]^2 * X[,5]))
}

Y = 5 + C + TrueH(X) + rnorm(n)

NLmod = NLint(Y=Y, X=X, C=C)

## Print posterior inclusion probabilities
NLmod$MainPIP

## Show the two way interaction probability matrix
NLmod$InteractionPIP

}

